% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_V}
\alias{get_V}
\alias{getV}
\title{Getting asymptotic variance of modified wilcox test statistic}
\usage{
getV(x_, y_ = NULL)
}
\arguments{
\item{x_}{A data set}

\item{y_}{Another data set, default is NULL}
}
\value{
Asymptotic variance of modified wilcoxon test statistic
}
\description{
Getting asymptotic variance of modified wilcox test statistic
}
