% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsymmtest.R
\name{mod.symm.test}
\alias{mod.symm.test}
\title{Modified symmetric test when the center of symmetry is unknown}
\usage{
mod.symm.test(
  x,
  y = NULL,
  alternative = c("two.sided", "less", "greater"),
  method = "wilcox"
)
}
\arguments{
\item{x}{data set to be tested}

\item{y}{another data set, default is NULL}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two sided" (default), "greater", or "less".
greater: test whether positively skewed (right-skewed),
less : test whether negatively skewed (left-skewed)}

\item{method}{a character string specifying which symmetric test to be used, "wilcox" is modified wilcoxon sign rank test,
and "sign" is modified sign test.}
}
\description{
Provides one unified symmetric test that incorporates two methods, "wilcox" and "sign". Also, paired data can be tested.
}
