% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{model}
\alias{model}
\title{Tree_model}
\format{
An rpart object that contains a tree model which fits alp to npilot, sk, kur, sd and distance_p2_half.
}
\usage{
model
}
\description{
Tree model object for wilcox method to find the optimal alp
}
\keyword{rpart}
