% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_model_sign.R
\docType{data}
\name{tree_model_sign}
\alias{tree_model_sign}
\title{Tree_model_sign}
\format{
An rpart object that contains a tree model which fits alp to npilot, sk, kur, sd.
}
\usage{
tree_model_sign
}
\description{
Tree model object to find the optimal alp for sign test method.
}
\keyword{rpart}
