% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsymmtest.R
\name{mod.symm.test}
\alias{mod.symm.test}
\title{Modified symmetry test when the center of symmetry is unknown}
\usage{
mod.symm.test(
  x,
  y = NULL,
  alternative = c("two.sided", "less", "greater"),
  method = "wilcox"
)
}
\arguments{
\item{x}{data set to be tested}

\item{y}{another data set, default is NULL}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two sided" (default), "greater", or "less".
greater: test whether positively skewed (right-skewed),
less : test whether negatively skewed (left-skewed)}

\item{method}{a character string specifying which symmetry test to be used, "wilcox" refers to wilcoxon signed rank test,
and "sign" is sign test.}
}
\value{
A list of class "htest" containing the following components:
  \itemize{
  \item statistic - the value of the test statistic.
  \item var - the asymptotic variance.
  \item alternative - a character string describing the alternative hypothesis.
  \item p.value - the p-value for the test.
  \item method - the type of test applied.

  }
}
\description{
Provides functionality that evaluate the effect of estimating the center of symmetry by sample mean
 on the signed-rank Wilcoxon test and sign test for symmetry. Also, paired data can be tested.
}
\details{
When "wilcox", the default method is used, the test statistic W is the statistic of signed-rank wilcox.test using sample mean 
as the center of symmetry; if the method "sign" is specified, the test statistic S would be the total number of the observations 
that smaller than sample mean. The asymptotic property of S could be found in the paper of Joseph Gastwirth.
}
\examples{
\dontrun{
  x <- rnorm(100, 1, 1)
  y <- rnorm(100)
  mod.symm.test(x)
  mod.symm.test(x,y)

}
}
\author{
Jiaojiao Zhou, Xinyu Gao
}
