% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_cal.R
\name{n.symm.test}
\alias{n.symm.test}
\title{Sample size determination for a nonparametric tests of symmetry}
\usage{
n.symm.test(
  x,
  sig.level = 0.05,
  power = 0.8,
  method = "wilcox",
  alternative = c("two.sided", "right.skewed", "left.skewed")
)
}
\arguments{
\item{x}{pilot data, numeric vector of data values.}

\item{sig.level}{significance level (Type I error probability)}

\item{power}{power of test (1 minus Type II error probability)}

\item{method}{a character string specifying which symmetry test to be used, "wilcox" refers to Wilcoxon signed-rank test,
and "sign" is sign test.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two sided" (default), "right.skewed", or "left.skewed".
You can specify just the initial letter.
"right.skewed": test whether positively skewed, 
"left.skewed" : test whether negatively skewed.}
}
\value{
A list of class "power.htest" containing the following components:
   \itemize{
  \item N - sample size estimated
  \item sig.level - significance level (Type I error probability)
  \item power - power of test (1 minus Type II error probability)
  \item method - the type of test applied
  \item alternative - one- or two-sided test. Can be abbreviated
  }
}
\description{
Determine the sample size required for one-sample wilcoxon signed-rank test or sign test 
with unknown center of symmetry estimated by sample mean.
}
\details{
Title
}
